refactor(sandbox): move rebuild gpu opt-out helper to focused module
